fix(storybook): move bar/race chart story to main bar stories
Close #1251
